ulebankoffice/src/main/java/es/unileon/ulebankoffice/service/UleBankEmployee.java
package es.unileon.ulebankoffice.service;

import org.hibernate.validator.constraints.NotBlank;


/**
 * @author Razvan Raducu
 *
 */
public class UleBankEmployee {
	
	@NotBlank
	private String userName;
	
	@NotBlank
	private String password;
	
	@NotBlank
	private String role;

	public String getUserName() {
		return userName;
	}

	public void setUserName(String userName) {
		this.userName = userName;
	}

	public String getPassword() {
		return password;
	}

	public void setPassword(String password) {
		this.password = password;
	}

	public String getRole() {
		return role;
	}

	public void setRole(String role) {
		this.role = role;
	}
	
	

}
